docGen/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "wm-docgen"
version = "0.1.0"
description = "Documentation generator for webMethods Integration Server packages."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "mkdocs>=1.5",
  "mkdocs-material>=9.5",
]

[project.scripts]
wm-docgen = "wm_docgen.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
